refactor: introduce service interfaces
