+++ src/main/java/com/willyes/clemenintegra/bom/service/DetalleFormulaService.java
@@ -1,32 +1,13 @@
 package com.willyes.clemenintegra.bom.service;
 
 import com.willyes.clemenintegra.bom.model.DetalleFormula;
-import com.willyes.clemenintegra.bom.repository.DetalleFormulaRepository;
-import org.springframework.stereotype.Service;
-import lombok.RequiredArgsConstructor;
 
 import java.util.List;
 import java.util.Optional;
 
-@Service
-@RequiredArgsConstructor
-public class DetalleFormulaService {
-
-    private final DetalleFormulaRepository detalleRepository;
-
-    public List<DetalleFormula> listarTodas() {
-        return detalleRepository.findAll();
-    }
-
-    public Optional<DetalleFormula> buscarPorId(Long id) {
-        return detalleRepository.findById(id);
-    }
-
-    public DetalleFormula guardar(DetalleFormula detalle) {
-        return detalleRepository.save(detalle);
-    }
-
-    public void eliminar(Long id) {
-        detalleRepository.deleteById(id);
-    }
+public interface DetalleFormulaService {
+    List<DetalleFormula> listarTodas();
+    Optional<DetalleFormula> buscarPorId(Long id);
+    DetalleFormula guardar(DetalleFormula detalle);
+    void eliminar(Long id);
 }
