+++ src/main/java/com/willyes/clemenintegra/bom/service/FormulaProductoService.java
@@ -1,33 +1,13 @@
 package com.willyes.clemenintegra.bom.service;
 
-import com.willyes.clemenintegra.bom.model.*;
-import com.willyes.clemenintegra.bom.repository.*;
-import org.springframework.stereotype.Service;
-import lombok.RequiredArgsConstructor;
+import com.willyes.clemenintegra.bom.model.FormulaProducto;
 
 import java.util.List;
 import java.util.Optional;
 
-@Service
-@RequiredArgsConstructor
-public class FormulaProductoService {
-
-    private final FormulaProductoRepository formulaRepository;
-
-    public List<FormulaProducto> listarTodas() {
-        return formulaRepository.findAll();
-    }
-
-    public Optional<FormulaProducto> buscarPorId(Long id) {
-        return formulaRepository.findById(id);
-    }
-
-    public FormulaProducto guardar(FormulaProducto formula) {
-        return formulaRepository.save(formula);
-    }
-
-    public void eliminar(Long id) {
-        formulaRepository.deleteById(id);
-    }
+public interface FormulaProductoService {
+    List<FormulaProducto> listarTodas();
+    Optional<FormulaProducto> buscarPorId(Long id);
+    FormulaProducto guardar(FormulaProducto formula);
+    void eliminar(Long id);
 }
-
