+++ src/main/java/com/willyes/clemenintegra/inventario/service/OrdenCompraDetalleService.java
@@ -1,33 +1,13 @@
 package com.willyes.clemenintegra.inventario.service;
 
 import com.willyes.clemenintegra.inventario.model.OrdenCompraDetalle;
-import com.willyes.clemenintegra.inventario.repository.OrdenCompraDetalleRepository;
-import org.springframework.stereotype.Service;
-import lombok.RequiredArgsConstructor;
 
 import java.util.List;
 import java.util.Optional;
 
-@Service
-@RequiredArgsConstructor
-public class OrdenCompraDetalleService {
-
-    private final OrdenCompraDetalleRepository repository;
-
-    public List<OrdenCompraDetalle> listarTodos() {
-        return repository.findAll();
-    }
-
-    public Optional<OrdenCompraDetalle> buscarPorId(Long id) {
-        return repository.findById(id);
-    }
-
-    public OrdenCompraDetalle guardar(OrdenCompraDetalle detalle) {
-        return repository.save(detalle);
-    }
-
-    public void eliminar(Long id) {
-        repository.deleteById(id);
-    }
+public interface OrdenCompraDetalleService {
+    List<OrdenCompraDetalle> listarTodos();
+    Optional<OrdenCompraDetalle> buscarPorId(Long id);
+    OrdenCompraDetalle guardar(OrdenCompraDetalle detalle);
+    void eliminar(Long id);
 }
-
