+++ src/main/java/com/willyes/clemenintegra/shared/security/service/JwtTokenServiceImpl.java
@@ -1,0 +1,49 @@
+package com.willyes.clemenintegra.shared.security.service;
+
+import com.willyes.clemenintegra.shared.model.Usuario;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import java.security.Key;
+import java.util.Date;
+
+@Service
+public class JwtTokenServiceImpl implements JwtTokenService {
+
+    private static final long EXPIRATION_MS = 3600_000; // 1 hora
+    private final Key secretKey;
+
+    public JwtTokenService(@Value("${clemen.jwt.secret}") String secret) {
+        this.secretKey = Keys.hmacShaKeyFor(secret.getBytes());
+    }
+
+    public String generarToken(Usuario usuario) {
+        return Jwts.builder()
+                .setSubject(usuario.getNombreUsuario())
+                .claim("rol", usuario.getRol().name())
+                .claim("usuarioId", usuario.getId())
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(System.currentTimeMillis() + EXPIRATION_MS))
+                .signWith(secretKey, SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    public Claims extraerClaims(String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(secretKey)
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    public String extraerNombreUsuario(String token) {
+        return extraerClaims(token).getSubject();
+    }
+
+
+}
+
